+++ modules/database/src/main/java/org/eclipse/xpanse/modules/database/register/DatabaseRegisterServiceStorage.java
@@ -4,7 +4,7 @@
  *
  */
 
-package org.eclipse.xpanse.orchestrator.register.impl;
+package org.eclipse.xpanse.modules.database.register;
 
 import jakarta.persistence.criteria.Predicate;
 import java.util.ArrayList;
@@ -13,10 +13,7 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 import org.apache.commons.lang3.StringUtils;
-import org.eclipse.xpanse.modules.database.register.RegisterServiceEntity;
-import org.eclipse.xpanse.modules.database.register.RegisterServiceRepository;
 import org.eclipse.xpanse.modules.models.query.RegisteredServiceQuery;
-import org.eclipse.xpanse.orchestrator.register.RegisterServiceStorage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Component;
