+++ modules/orchestrator/src/main/java/org/eclipse/xpanse/orchestrator/OrchestratorPlugin.java
@@ -7,23 +7,16 @@
 package org.eclipse.xpanse.orchestrator;
 
 import org.eclipse.xpanse.modules.credential.AuthenticationCapabilities;
-import org.eclipse.xpanse.modules.deployment.DeployResourceHandler;
 import org.eclipse.xpanse.modules.models.enums.Csp;
 import org.eclipse.xpanse.modules.monitor.MetricsExporter;
+import org.eclipse.xpanse.orchestrator.service.ServiceResourceHandler;
 
 /**
  * This interface describes orchestrator plugin in charge of interacting with backend fundamental
  * APIs.
  */
-public interface OrchestratorPlugin extends AuthenticationCapabilities, MetricsExporter {
-
-    /**
-     * get the resource handler of the CSP plugin.
-     *
-     * @return the resource handler of the plugin.
-     */
-    DeployResourceHandler getResourceHandler();
-
+public interface OrchestratorPlugin
+        extends ServiceResourceHandler, AuthenticationCapabilities, MetricsExporter {
 
     /**
      * get the Csp of the plugin.
